refactor: 애니메, 시리즈 양방향 매핑 #3
